add TokenDivider
